pythonProject61/weather app/requirements.txt
alabaster==0.7.12
altair==5.4.1
appdirs==1.4.4
arabic-reshaper==2.1.3
astunparse==1.6.3
atomicwrites==1.4.0
attrs==24.3.0
av==10.0.0
Babel==2.9.1
badapted==0.0.3
beautifulsoup4==4.11.1
blinker==1.8.2
blosc2==2.0.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.8
codecov==2.1.12
colorama==0.4.4
coloredlogs==15.0.1
coverage==6.1.2
cryptography==35.0.0
ctranslate2==3.18.0
cycler==0.11.0
Cython==0.29.24
darc-experiment-toolbox==0.0.1
decorator==4.4.2
distlib==0.3.6
docutils==0.17.1
dukpy==0.2.3
egi-pynetstation==0.0.0rc1
elementpath==4.1.5
esprima==4.0.1
et-xmlfile==1.1.0
faster-whisper==0.7.1
ffpyplayer==4.3.5
filelock==3.9.0
flatbuffers==23.5.26
freetype-py==2.2.0
fsspec==2023.6.0
ftd2xx==1.1.2
future==0.18.2
gevent==21.8.0
gitdb==4.0.9
GitPython==3.1.24
glfw==2.4.0
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-speech==2.11.1
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-status==1.41.1
html2text==2020.1.16
httplib2==0.20.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.3
imageio==2.10.3
imageio-ffmpeg==0.4.5
imagesize==1.3.0
importlib_resources==6.4.5
javascripthon==0.12
jedi==0.18.0
Jinja2==3.0.3
json-tricks==3.15.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.3.2
LabJackPython==2.0.4
libcst==0.3.21
macropy3==1.1.0b2
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.4.3
mdurl==0.1.2
more-itertools==8.11.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.2
msgpack-numpy==0.4.7.1
mypy-extensions==0.4.3
narwhals==1.21.1
ntplib==0.4.0
numexpr==2.7.3
numpy==1.24.4
onnxruntime==1.15.1
opencv-python==4.5.4.58
openpyxl==3.0.9
packaging==21.2
pandas==2.0.3
parso==0.8.2
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
pluggy==1.0.0
pocketsphinx==0.1.15
proglog==0.1.9
proto-plus==1.19.8
protobuf==5.29.2
psutil==5.8.0
psychopy @ file:///C:/Users/lpzjwp/code/psychopy/git
psychtoolbox @ file:///G:/Shared%20drives/Dev/py/libs/psychtoolbox/psychtoolbox-3.0.18.2-cp38-cp38-win_amd64.whl
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.9.1
pygame==2.1.0
pyglet==1.4.11
Pygments==2.19.1
pylsl==1.15.0
pyo==1.0.4
PyOpenGL==3.1.5
pyparallel==0.2.2
pyparsing==2.4.7
pypi-search==1.2.1
pypiwin32==223
pyqmix==2021.1.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyreadline3==3.4.1
pyserial==3.5
pytest==4.0.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-gitlab==2.10.1
python-pygaze==0.7.3
python-vlc==3.0.11115
pytz==2021.3
pywin32==302
pyWinhook==1.6.2
pyxid2==1.0.3
PyYAML==6.0
pyzmq==22.3.0
questplus==2023.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.9.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.7.2
ruamel.yaml==0.16.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
snowballstemmer==2.1.0
sounddevice==0.4.3
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sr-research-pylink==2.1.1.0
streamlit==1.40.1
sympy==1.12
tables==3.8.0
tenacity==9.0.0
Titta @ git+https://github.com/marcus-nystrom/Titta.git@07772b4883f3dbd97739c9e6693bb25a170aeb52
tobii-research==1.10.1
tokenizers==0.13.3
toml==0.10.2
tornado==6.4.2
tqdm==4.62.3
typing-inspect==0.7.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==4.2.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.19.0
watchdog==4.0.2
websocket-client==1.2.1
websockets==11.0.3
winrt==1.0.21033.1
wxPython==4.1.1
xarray==0.20.1
xlrd==2.0.1
xmlschema==2.4.0
zipp==3.20.2
zope.event==4.5.0
zope.interface==5.4.0
requests==2.26.0
